change BigDecimal to int

+++ src/com/doaa/vetclinic/DAOs/ClinicDAO.java
@@ -3,7 +3,6 @@
  */
 package com.doaa.vetclinic.DAOs;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import com.doaa.vetclinic.entities.Clinic;
@@ -16,13 +15,13 @@ public interface ClinicDAO {
 
 	public List<Clinic> listAllClinics();
 	
-	public List<Clinic> searchClinicByPhone(BigDecimal phoneNum);
+	public List<Clinic> searchClinicByPhone(int phoneNum);
 	
 	public List<Clinic> searchClinicByAddress(String address);
 
 	public void saveClinic(Clinic clinic);
 
-	public void deleteClinic(BigDecimal id);
+	public void deleteClinic(int id);
 	
-	public Clinic getClinicById(BigDecimal id);
+	public Clinic getClinicById(int id);
 }
